feat: V4 — ondalik AI skor (8.7, 6.3 gibi) + gelismis prompt
- ai_score INTEGER → FLOAT (DB migration v25)
- Skor hassasiyeti: 1.0-10.0 arasi, 0.1 adimlarla
- Gelismis prompt: ihale tutari, kar artis yuzdesine gore farkli puanlar
- %100 artis vs %32 artis farkli puan alir
- TradingView icerik + gpt-4o analiz (V3 devami)
- Tweet + admin mesajlari ondalik format (8.7/10)
- Frontend ChatBubble ondalik gosterim

+++ app/schemas.py
@@ -230,7 +230,7 @@ class TelegramNewsOut(BaseModel):
     gap_pct: Optional[Decimal] = None
     prev_close_price: Optional[Decimal] = None
     theoretical_open: Optional[Decimal] = None
-    ai_score: Optional[int] = None
+    ai_score: Optional[float] = None
     ai_summary: Optional[str] = None
     kap_url: Optional[str] = None
     message_date: Optional[datetime] = None
